add extended comments on all execpt utilities
