+++ lib/components/displayComponents/buttons/alphabets.dart
@@ -1,12 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
+///Stateful Aplphabet handler that is initialised for use by
+///calling [Alphabet().createState()]
+///The [createState()] adds an awareness of the widget state to
+///the widget tree and enables overlaying the [AlphabetButton] widget
+///it serves as the [alphabetHandler] for the [EntryHandler()] widget.
+
 class Alphabet extends StatefulWidget {
   @override
   _AlphabetState createState() => _AlphabetState();
 }
 
 class _AlphabetState extends State<Alphabet> {
+  ///Alphabet List that contains all the User triggered entries
   List<String> newAlpha = [];
   String entry;
 
@@ -26,6 +33,11 @@ class _AlphabetState extends State<Alphabet> {
     );
   }
 
+  ///combine the [String] alphabets contained in the [newAlpha]
+  ///into one concantenated String
+  ///
+  ///A [newAlpha] of ['a','c','u','t','e'] returns a value of
+  /// 'acute'
   String allAlphabets() {
     String alpha = '';
 
@@ -37,15 +49,28 @@ class _AlphabetState extends State<Alphabet> {
     return alpha;
   }
 
+  ///reset the [newAlpha] to an empty [List<String>]
   void reset() {
     newAlpha = [];
   }
 }
 
+///[AlphabetButton] creates a [Card] that is wrapped with
+///a [GestureDetector] it is usually used in conjunction with a [LetterMap]
+///and an [Alphabet] state handler
 class AlphabetButton extends StatelessWidget {
   AlphabetButton({this.alphabet, this.active, this.onPressed});
+
+  /// [bool] active manages the state of the UI of the button
+  /// if the button has been triggered the button color is transformed
+  /// to a completely transparent [Card] with a minor elevation backdrop.
   final bool active;
+
+  /// [String] alphabet: is the child of the [AlphabetButton]
+  /// the actual alphabet to be displayed
   final String alphabet;
+
+  /// [Function] a function to be triggered on the button press detected.
   final Function onPressed;
 
   @override
