+++ lib/components/displayComponents/buttons/slimButtons.dart
@@ -2,6 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:wordsmith/utilities/constants.dart';
 
 class SlimButton extends StatelessWidget {
+  ///returns a Material Button that takes a
+  ///by default a [Text] as a child but can be
+  ///instructed to take any kind of widget as a child
+  ///when the [useWidget] argument is set as true
+  ///this becomes useful if a loading indicator is desired
+  ///to be shown in the place of the text when a
+  ///pre-specified condition has been met
+  ///of course a setState is called outside the [SlimButton]
+  ///scope to enable it rebuild it appearance
   SlimButton(
       {this.onTap,
       this.label,
@@ -9,11 +18,25 @@ class SlimButton extends StatelessWidget {
       this.textColor,
       @required this.useWidget,
       this.widget});
+
+  ///label text of the custom [Material] button
   final String label;
+
+  /// [Function] to be performed when a tap gesture is detected
   final Function onTap;
+
+  /// basic color of the  button area defaults to material [Colors.white10]
   final Color color;
+
+  ///Color of the text defaults to material [Colors.black]
   final Color textColor;
+
+  /// [bool] indicates whether the material button should accept
+  /// any widget other than a Text widget as a child
   final bool useWidget;
+
+  /// Secondary widget that is only displayed when the useWiget
+  ///  Value is set to true
   final Widget widget;
 
   @override
@@ -29,7 +52,8 @@ class SlimButton extends StatelessWidget {
                 : Text(
                     label,
                     style: TextStyle(
-                        color: this.textColor, fontWeight: FontWeight.w600),
+                        color: textColor ?? Colors.black,
+                        fontWeight: FontWeight.w600),
                   ),
           ),
         ),
@@ -37,16 +61,22 @@ class SlimButton extends StatelessWidget {
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(20),
         ),
-        color: color);
+        color: color ?? Colors.white10);
   }
 }
 
-
-
-
 class SignUp extends StatelessWidget {
-  const SignUp({Key key, @required this.routeName,@required this.label}) : super(key: key);
+  ///A relatively tiny button that tightly hugs its text child
+  ///Used in minor registration navigation
+  ///accepts [String] routeName as registered on the [main.dart] file
+  ///and a label
+  const SignUp({Key key, @required this.routeName, @required this.label})
+      : super(key: key);
+
+  ///routeName as registered on the main dart file
   final String routeName;
+
+  ///label text on the button
   final String label;
 
   @override
@@ -56,7 +86,7 @@ class SignUp extends StatelessWidget {
       color: Colors.white.withOpacity(.1),
       child: InkWell(
         borderRadius: BorderRadius.circular(10),
-        onTap: () => Navigator.pushReplacementNamed(context,routeName),
+        onTap: () => Navigator.pushReplacementNamed(context, routeName),
         child: Padding(
           padding: const EdgeInsets.all(8.0),
           child: Text(
